scoped configuration management
